refactor: change package name to bluetoothingspeaker
