+++ audio-sink/src/main/AndroidManifest.xml
@@ -24,9 +24,10 @@
     <uses-permission android:name="com.google.android.things.permission.USE_PERIPHERAL_IO" />
     <application android:label="@string/app_name">
         <uses-library android:name="com.google.android.things"/>
-        <activity android:name="io.github.rosariopfernandes.btspeaksthings.A2dpSinkActivity">
+        <activity android:name="io.github.rosariopfernandes.bluetoothingspeaker.A2dpSinkActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.HOME"/>
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
