+++ audio-sink/src/main/java/io/github/rosariopfernandes/bluetoothingspeaker/A2dpSinkActivity.kt
@@ -1,4 +1,4 @@
-package io.github.rosariopfernandes.btspeaksthings
+package io.github.rosariopfernandes.bluetoothingspeaker
 
 import android.app.Activity
 import android.bluetooth.BluetoothAdapter
