+++ audio-sink/src/main/java/io/github/rosariopfernandes/bluetoothingspeaker/Extensions.kt
@@ -1,4 +1,4 @@
-package io.github.rosariopfernandes.btspeaksthings
+package io.github.rosariopfernandes.bluetoothingspeaker
 
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothDevice
