+++ audio-sink/src/main/res/layout/activity_a2dpsink.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="io.github.rosariopfernandes.btspeaksthings.A2dpSinkActivity">
+        tools:context="io.github.rosariopfernandes.bluetoothingspeaker.A2dpSinkActivity">
 
     <TextView
         android:id="@+id/textView"
